test(checkout): CHECKOUT-9170 Combine files
